RangeProof::CreatorParams - generalized to save/recover arbitrary data

+++ core/shielded.h
@@ -81,7 +81,7 @@ namespace beam
 		protected:
 			void GenerateInternal(ShieldedTxo&, const ECC::Hash::Value& hvShared, ECC::Oracle&, Key::IPKdf& gen);
 			static void get_Seed(ECC::uintBig&, const ECC::Hash::Value& hvShared);
-			static uint32_t Msg2Scalar(ECC::Scalar::Native&, const ECC::uintBig&);
+			static uint8_t Msg2Scalar(ECC::Scalar::Native&, const ECC::uintBig&);
 			static void Scalar2Msg(ECC::uintBig&, const ECC::Scalar::Native&, uint32_t);
 		};
